refactor: API 엔드포인트에 memberID 추가 및 controller, service 위치 변경
