+++ src/main/java/com/example/muaring/domain/social/service/PostService.java
@@ -123,24 +123,6 @@ public class PostService {
         groupPlaylistRepository.save(groupPlaylist);
     }
 
-    @Transactional
-    public Page<MusicHistoryDTO> getMusicHistoryByMember(Long memberId, Integer year, Integer month, Pageable pageable) {
-        if (!memberRepository.existsById(memberId)) {
-            throw new MusicException(MusicErrorCode.MEMBER_NOT_FOUND);
-        }
-
-        Page<MusicPost> posts = musicPostRepository.findByMemberAndYearMonth(memberId, year, month, pageable);
-
-        return posts.map(post -> MusicHistoryDTO.builder()
-                .postId(post.getId())
-                .musicId(post.getMusic().getId())
-                .title(post.getMusic().getName())
-                .artist(post.getMusic().getArtistName())
-                .albumImage(post.getMusic().getAlbumImgUrl())
-                .createdAt(post.getCreatedAt())
-                .build());
-    }
-
     @Transactional(readOnly = true)
     public Page<MusicPostFeedResponseDto> getTodayFolloweePosts(Pageable pageable) {
